Add endpoint topology-based filtering
In order to implement service topology, the destination service needs to filter all
available endpoints of a service based on its topological preference. This PR introduces
endpoint filtering in the translator component of the destination service. To achieve this,
the endpoint translator (listener) has now been made stateful. Based on the topology pref
of a service, its endpoints are filtered; updates are propagated through a diff-based API
on the translator side, similar to how addresses are being handled in the EndpointsWatcher.

Fixes #4503

Signed-off-by: Matei David <matei.david.35@gmail.com>