move public-api into a add-on
This PR movies the public-api component into its own add-on, thus making
it optional.

This follows the same model as other add-ons, except that
it defines its own struct i.e `publicApi` with fields rather than being
a `map[string]interface{}` so that these fields can be used in code to
be able to be overriden from options.

Signed-off-by: Tarun Pothulapati <tarunpothulapati@outlook.com>