mount linkerd-config-addons to web
This commit updates linkerd-web component to load linkerd-config-addons
cm as a VolumeMount. web then uses the config to check enabled add-ons
for grafana and jaeger and sets the addr variables based on that.

This has to be done this way, as there is no way to get values
configuration between add-ons i.e grafana and jaeger values into web
add-on.

Signed-off-by: Tarun Pothulapati <tarunpothulapati@outlook.com>