Add missing install flag when DisableHeartBeat value is provided
Problem
I am getting the following errors when using helm with `DisableHeartBeat` set to `true`
```
linkerd endpoints <endpoint>
Cannot find Linkerd: missing ServiceAccounts: linkerd-heartbeat
Validate the install with: linkerd check
```
It is also failing `check`.
```
× heartbeat ServiceAccount exist
    missing ServiceAccounts: linkerd-heartbeat
    see https://linkerd.io/checks/#l5d-existence-sa for hints
```
This issue was introduced here:
https://github.com/linkerd/linkerd2/issues/3386
The command was added to the install command, but not included in the helm chart.

Solution is to include it in the helm chart options.

Once merged, it should no longer return the error and return a successful `check` result.

Fixes #4834

Signed-off-by: mvaal <mjvaal@gmail.com>