Integration test for smi-metrics
This PR adds a integration test which installs SMI-Metrics and performs
queries and matches the reply with a regex query.

Currently, As the smi-metrics helm chart is not published publicly it is
stored locally but publishing is currently being worked on and this
PR will be updated to pull that chart before being merged.

Signed-off-by: Tarun Pothulapati <tarunpothulapati@outlook.com>