Allow external redirection for Grafana and jaeger
Fixes #4838

Currently, For Grafana and Jaeger, the web component acts as a proxy to
the given grafana isntance (could be a external one too, using
global.grafanaURL). For proxying, Grafana and Jaeger have to be
configured to allow requests from a proxy and this also this makes
grafana not accept requests for normal URLs. When a user is using a
external grafana instance, this defy's the purpose as the external
grafana instance can't accept normal requests.

This PR updates the web component with a new flag called `--redirect`
which when configured redirects instead of proxying. The same flag will
be used both for grafana and jaeger.

Signed-off-by: Tarun Pothulapati <tarunpothulapati@outlook.com>