Fix cli upgrade for endpoint slice feature gate
When upgrading the linkerd control plane through the cli, if upgrading
with endpoint slices turned on (when they were previously off), the new flag
does not persist in the configmap. To fix this, the configmap value should now be updated
if upgrading with slices turned on. Additionally, the helm configmap tmpl also
used a different name to the protobuf file, this has been changed to avoid future issues.

Signed-off-by: Matei David <matei.david.35@gmail.com>