update upgrade flow to convert protobuf Duration into golang Duration
Protobuf's Duration is different from goland Duration, and hence this
conversion has to be done manually during retrieval. This worked
previously because we were not using that value directly, and were
storing the same value in protobuf format once again (i.e in the
configMap) but now with these duration values being used as flags outside
the configmap, this was discovered.

Signed-off-by: Tarun Pothulapati <tarunpothulapati@outlook.com>