Update to Go 1.15
Go 1.15 has been released recently, and projects like Kubernetes, etc
have already moved to it. This PR updates the go versions in the repo,
and the docker images.

Updating only the go version, makes the identity fail because of go
updating how it precieves certificates. Code has to be updated in cert
generation to make sure theere is SAN fields, with the value of that of
common field.

Signed-off-by: Tarun Pothulapati <tarunpothulapati@outlook.com>