Use latest config secret for CLI functionality
This commit updates the current CLI functions to use the latest secret
resource if present or fall back to the older configuration. This is done
by replacing the FetchLinkerdConfigMap with a newer FetchLinkerdConfiugration
which returns Values, and is used to know the cluster state.

This also moves the mergeRaw functionality outside to the values, to be reused

Signed-off-by: Tarun Pothulapati <tarunpothulapati@outlook.com>