default install of linkerd2 interferes with apiserver reliability
The default installation of linkerd2 uses an opt-out model. If you
consider an existing cluster with 1000s of pods and 100s of nodes, and
add linkerd2 without customizing coniguration, it's very likely to
impact cluster reliability.

This PR adds objectSelector from k8s 1.15+ that moves the
linkerd.io/inject selector from the apiservice to the apiserver
configuration. This is an important when the linkerd2 proxy injector
is down, the pods that have no interest in linkerd will not be
prevented from pod creation.

This configuration is almost identical to istio's istio/istio#16539

Signed-off-by: Drew Wells <drew.wells00@gmail.com>