Decode durations from protobuf into golang correctly
This PR updates the retrieval of durations from linkerd-config
to have conversion into the normal Go duration. This is needed
as  the duration types in protobuf and goland are different.

This did not break anything previously, as we were directly using
this fields only for protobuf generation for upgrades, but now that
we are trying to use this fields with the `Values` go types (with the
new config format) this causes the ugprade tests to fail.

Signed-off-by: Tarun Pothulapati <tarunpothulapati@outlook.com>