Restrict controlPlaneTracing field only to control plane components
Previously, `global.controlPlaneTracing` was not available during
injection and thus not effecting it.

But now that it is stored and retrieved, It causes problems when
`global.controlPlaneTracing` is enabled while no `trace-collector` value
is present or specified(through flags).

This can be further be simplified by removing `controlPlaneTracing` all
together and have the trace flags under `install` command, and this
effort could be part of the minimal control plane effort.

Signed-off-by: Tarun Pothulapati <tarunpothulapati@outlook.com>