cli: make check return SkipError when there is no prometheus configured
Fixes #5143

Check currently keeps retrying for a bit when there is no prometheus
instnace configured i.e both linkerd and external prometheus instance
(through global.prometheusUrl).

The availability of prometheus is useful for some calls in public-api
that the check uses. This change updates `hc.getDataPlanePods` to return
a `SkipError` if its a prometheus error.

Signed-off-by: Tarun Pothulapati <tarunpothulapati@outlook.com>