upgrade: set identity.issuer.crtExpiry correctly with legacy upgrades
With legacy upgrades, we can parse the cert and store the expiry
correctly instead of storing it as the default value which could be a
problem when we use that field. Currently, we do not use this field and
hence it did not cause any problems.

With Install on the latest edges, This field is correctly set and works
as expected. Thus, upgrades also hav the right value.

Signed-off-by: Tarun Pothulapati <tarunpothulapati@outlook.com>