proxy: v2.120.0
This release modifies the outbound proxy to bypass service discovery and
load balancing when the control plane does not provide a resolveable
service name. The control plane may optionally provide endpoint metadata
so that mTLS can be established, metrics can be labeled, etc, without
further discovery.

HTTP/2 clients and servers are now configured to send PING messages to
validate that the peer application is able to serve requests.

---

* h2: enable HTTP/2 keepalive PING frames (linkerd/linkerd2-proxy#737)
* actions: Add timeouts to GitHub actions (linkerd/linkerd2-proxy#738)
* outbound: Skip endpoint resolution on profile hint (linkerd/linkerd2-proxy#736)