cli: don't check for SAN in root and intermediate Certs
Fixes #5228

As discussed in the issue, It is not correct for root and intermediate
certs to have SAN. This PR updates the check to not verify the
intermediate issuer cert with the identity dns name (which checks with
SAN and not CN as the the `verify` func is used to verify leaf certs and
not root and intermediate certs). This PR also removes the extra san
field added in cert generation in CLI which was added previously.

Signed-off-by: Tarun Pothulapati <tarunpothulapati@outlook.com>